Cleanups and comments
