+++ ctapipe/coordinates/coordinate_base.py
@@ -1,3 +1,6 @@
+"""
+
+"""
 import numpy as np
 
 __all__ = [
@@ -6,14 +9,28 @@ __all__ = [
 
 
 class BaseCoordinate:
+    """
 
+    """
+
+    #
     system_order, transformations, reverse_transformations = None, None, None
 
     def init(self):
         return
 
     def _find_path(self, new_system):
+        """
 
+        Parameters
+        ----------
+        self
+        new_system
+
+        Returns
+        -------
+
+        """
         path_start = np.where(self.system_order == self.__class__.__name__)[0][0]
         path_end = np.where(self.system_order == new_system)[0][0]
 
@@ -23,6 +40,17 @@ class BaseCoordinate:
             return reversed(self.reverse_transformations[path_start:path_end])
 
     def transform_to(self, new_system):
+        """
+
+        Parameters
+        ----------
+        self
+        new_system
+
+        Returns
+        -------
+
+        """
 
         transform_path = self._find_path(new_system)
